# Conflicts:
#	src/main/java/de/mpg/biochem/mars/image/commands/BeamProfileCorrectionCommand.java

+++ src/main/java/de/mpg/biochem/mars/metadata/MarsOMEChannel.java
@@ -164,7 +164,7 @@ public class MarsOMEChannel extends AbstractJsonConvertibleRecord {
 		});
 
 		setJsonField("gain", jGenerator -> {
-			if (gain != null) jGenerator.writeNumberField("gain", gain.doubleValue());
+			if (gain != null) jGenerator.writeNumberField("gain", gain);
 		}, jParser -> gain = jParser.getDoubleValue());
 
 		// Should we keep track of the units here ???
